+++ src/main/java/mycompany/sebastian/portfolioBackend/Repository/UserDataRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserDataRepository extends JpaRepository<UserData, Long> {
+public interface UserDataRepository extends JpaRepository<UserData, Integer> {
     UserData findByUserName(String userName);
     Optional<UserData> findById(int id);
 }
